refactor(verifiers): split shared validation support
