+++ benchmarks/validation/mathematical_benchmarks_v1/test_fiber_dimension_semicontinuity_repair.py
@@ -164,8 +164,6 @@ def test_oversized_fiber_list_rejected_without_crash(tmp_path: Path) -> None:
 
 def test_task_metadata_declares_diagnostics() -> None:
     """Diagnostics are declared in task-local metadata, not global registries."""
-    assert TASK not in support.INPUT_BINDING_DECOUPLED_TASKS
-    assert TASK not in support.SCOPE_INDEPENDENT_ASSURANCE_TASKS
     assert support.is_input_binding_decoupled(TASK) is True
     assert support.is_scope_independent_assurance(TASK) is True
     metadata = support.load_task_contract_metadata(TASK)
